refactor package of models
