+++ gotour/src/main/java/InitDatabase.java
@@ -1,5 +1,4 @@
 
-import models.*;
 import org.hibernate.SessionFactory;
 import org.hibernate.cfg.AnnotationConfiguration;
